// packages/api/src/modules/dataviz/amounts-vs-program-region/entities/AmountsVsProgramRegionEntity.ts
type AmountsVsProgramRegionEntity = {
    // For the moment this concerns only chorus data

    exerciceBudgetaire: number;
    programme: string;
    mission: string | null;
    montant: number;
    // region can contain data about central administration, not only regions
    regionAttachementComptable: string;
};

export default AmountsVsProgramRegionEntity;